Organize k8s data types in a different way
