+++ src/model/services/mapi/capiv1alpha3/getClusterList.ts
@@ -2,7 +2,7 @@ import { HttpRequestMethods, IHttpClient } from 'model/clients/HttpClient';
 import * as k8sUrl from 'model/services/mapi/k8sUrl';
 import { LoggedInUserTypes } from 'stores/main/types';
 
-import * as capiv1alpha3 from './types/capiv1alpha3';
+import { IClusterList } from './';
 
 export function getClusterList(client: IHttpClient, user: ILoggedInUser) {
   return async () => {
@@ -17,7 +17,7 @@ export function getClusterList(client: IHttpClient, user: ILoggedInUser) {
     client.setRequestMethod(HttpRequestMethods.GET);
     client.setAuthorizationToken(user.auth.scheme, user.auth.token);
 
-    const response = await client.execute<capiv1alpha3.IClusterList>();
+    const response = await client.execute<IClusterList>();
 
     return response.data;
   };
