+++ src/model/services/mapi/capiv1alpha3/types.ts
@@ -1,5 +1,5 @@
-import * as corev1 from '../../types/corev1';
-import * as metav1 from '../../types/metav1';
+import * as corev1 from '../corev1';
+import * as metav1 from '../metav1';
 
 export interface IClusterNetwork {
   apiServerPort: number;
