+++ src/model/services/mapi/corev1/index.ts
@@ -1,0 +1,1 @@
+export * from './types';
