+++ src/model/services/mapi/corev1/types.ts
